OLLib: change directory structure
In preparation to pull in lilyglyphs and musicexamples I change directory structure.
This commit places all OLLib components inside one /OLLib directory.
This file will be the one that the LilyPond user will have to put in the include path.
Then he can write:
\include "OLLib.ily"
or
\include "OLLincludes/..."
which looks like a reasonable syntax for the actual documents